+++ packages/ui/voo_forms/lib/src/presentation/atoms/fields/voo_radio_field_widget.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:voo_forms/src/domain/entities/form_field.dart';
+import 'package:voo_forms/src/presentation/widgets/voo_field_options.dart';
 import 'package:voo_ui_core/voo_ui_core.dart';
 
 class VooRadioFieldWidget<T> extends StatelessWidget {
   final VooFormField<T> field;
+  final VooFieldOptions options;
   final ValueChanged<T?>? onChanged;
   final String? error;
   final bool showError;
@@ -11,6 +13,7 @@ class VooRadioFieldWidget<T> extends StatelessWidget {
   const VooRadioFieldWidget({
     super.key,
     required this.field,
+    required this.options,
     this.onChanged,
     this.error,
     this.showError = true,